# job-portal-api/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Job Portal REST API",
    "description": "Comprehensive REST API documentation for the Job Portal system supporting Candidates, Companies, Jobs, Applications, Messaging, and Shortlisting.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local Development Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your JWT token in the format: Bearer <token>"
      }
    },
    "schemas": {
...
      "JobVacancyRequest": {
        "type": "object",
        "required": [
          "title",
          "employmentTypeId",
          "workplaceTypeId"
        ],
        "properties": {
          "title": { "type": "string", "example": "Senior React Developer" },
          "description": { "type": "string", "example": "React.js development role." },
          "employmentTypeId": { "type": "integer", "example": 1, "description": "1: Full-time, 2: Part-time, 3: Contract, 4: Internship" },
          "salary": { "type": "number", "example": 120000.00 },
          "currency": { "type": "string", "example": "INR" },
          "countryId": { "type": "integer", "example": 1 },
          "workplaceTypeId": { "type": "integer", "example": 1, "description": "1: Remote, 2: On-site, 3: Hybrid" },
          "requiredSkills": {
            "type": "array",
            "items": { "type": "string" },
            "example": ["React", "Redux", "TypeScript"]
          },
          "deadline": { "type": "string", "format": "date", "example": "2025-09-30" },
          "status": { "type": "string", "enum": ["open", "closed"], "example": "open" }
        }
      },
...
      "MessageRequest": {
        "type": "object",
        "required": [
          "receiver_id",
          "body"
        ],
        "properties": {
          "receiver_id": { "type": "string", "format": "uuid", "example": "11111111-1111-1111-1111-111111111111" },
          "body": { "type": "string", "example": "Hello, I am interested in your open positions." }
        }
      }
    }
  },
  "paths": {
...
    "/api/jobs": {
      "get": {
        "tags": ["Jobs"],
        "summary": "Get all active jobs (with filters)",
        "parameters": [
          { "name": "search", "in": "query", "schema": { "type": "string" }, "description": "Search in title/description" },
          { "name": "countryId", "in": "query", "schema": { "type": "integer" } },
          { "name": "workplaceTypeId", "in": "query", "schema": { "type": "integer" } },
          { "name": "employmentTypeId", "in": "query", "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": { "type": "array", "items": { "type": "object" } }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Jobs"],
        "summary": "Create job vacancy (company only)",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobVacancyRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Job vacancy created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "message": { "type": "string", "example": "Job vacancy created successfully" },
                    "data": { "type": "object" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/jobs/{id}": {
      "get": {
        "tags": ["Jobs"],
        "summary": "Get single job detail",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" },
            "example": "11111111-1111-1111-1111-111111111114"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": { "type": "object" }
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Jobs"],
        "summary": "Update job vacancy (company owner only)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" },
            "example": "11111111-1111-1111-1111-111111111114"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobVacancyRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Job vacancy updated successfully"
          }
        }
      },
      "delete": {
        "tags": ["Jobs"],
        "summary": "Delete job vacancy (company owner only)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" },
            "example": "11111111-1111-1111-1111-111111111114"
          }
        ],
        "responses": {
          "200": {
            "description": "Job vacancy deleted successfully"
          }
        }
      }
    },
...
    "/api/messages": {
      "post": {
        "tags": ["Messages"],
        "summary": "Send a message to a user",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MessageRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Message sent successfully"
          }
        }
      }
    },
    "/api/messages/conversations": {
      "get": {
        "tags": ["Messages"],
        "summary": "Get all conversations (active chats) for logged-in user",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": { "type": "array", "items": { "type": "object" } }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/messages/user/{userId}": {
      "get": {
        "tags": ["Messages"],
        "summary": "Get message history with a specific user",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" },
            "example": "11111111-1111-1111-1111-111111111111"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": { "type": "array", "items": { "type": "object" } }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}